requirements.txt
Django>=4.0.3,<4.1
djangorestframework>=2.4.8,<4.5
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,>0.23
Pillow>=9.2.0,<9.3
uwsgi>=2.0.20,<2.0.21